requirements.txt
# === Основные библиотеки ===
numpy==1.24.4
pandas==1.5.3
scikit-learn==1.2.2
catboost==1.2
matplotlib==3.6.3

# === Технический анализ ===
ta==0.10.2
TA-Lib==0.4.28  # Устанавливается вручную через .whl под Python 3.10

# === Обработка дисбаланса ===
imbalanced-learn==0.10.1

# === Байесовская оптимизация ===
bayesian-optimization==1.4.3
colorama==0.4.6

# === Pybit SDK для Bybit Unified API ===
pybit==5.3.0

# === Прочее ===
tqdm==4.66.1
graphviz==0.20.3
plotly==6.1.1
joblib==1.5.1
threadpoolctl==3.6.0
python-dateutil==2.9.0.post0
pillow==11.2.1     # при проблемах с matplotlib попробуй 10.3.0
pyparsing==3.2.3
kiwisolver==1.4.8
fonttools==4.58.0
cycler==0.12.1
contourpy==1.3.2

# === ENV ===
python-dotenv==1.0.1
